Create Frontend.Fragment.{IdeaList, QuorumBar} modules.
