+++ src/Frontend/Page/Topic.hs
@@ -23,7 +23,7 @@ where
 import Action (ActionM, ActionPersist(..), ActionUserHandler, getCurrentTimestamp)
 import Control.Exception (assert)
 import Frontend.Fragment.Category
-import Frontend.Page.Overview
+import Frontend.Fragment.IdeaList
 import Frontend.Prelude hiding (moveIdeasToLocation, editTopic)
 
 import qualified Action (createTopic)
