+++ tests/AllModulesSpec.hs
@@ -21,6 +21,8 @@ import Frontend ()
 import Frontend.Core ()
 import Frontend.Fragment.Category ()
 import Frontend.Fragment.Comment ()
+import Frontend.Fragment.IdeaList ()
+import Frontend.Fragment.QuorumBar ()
 import Frontend.Page ()
 import Frontend.Page.Admin ()
 import Frontend.Page.Delegation ()
